Re-add and debug all views withdrawn in the previous reset
There were issues that needed addressing, particularly regarding a chain reaction of errors
related to the fact that the slug is used to identify the reference object 'Dream Job' as
the basis for matching calculation. I needed to ensure that users cannot call views requiring
the slug before that object is created. Therefore, I have reset the commits to before the
introduction of the views and introduced them one by one, debugging them along the way.